% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_sort.R
\name{quick_sort}
\alias{quick_sort}
\title{Sort a numeric vector using Quick Sort}
\usage{
quick_sort(arr)
}
\arguments{
\item{arr}{A numeric vector to be sorted.}
}
\value{
A sorted numeric vector.
}
\description{
Sort a numeric vector using Quick Sort
}
\examples{
my_vec <- c(34, 7, 23, 32, 5, 62, 32)
quick_sort(my_vec)
}
